requirements.txt
flask
pyngrok
werkzeug
flask-cors
openai
pymongo[srv]
python-dotenv